ТЗ-22 Наполнены модули Delivery и Order

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/feignclient/PaymentClient.java
@@ -21,7 +21,7 @@ public interface PaymentClient {
     void paymentSuccess(@RequestBody UUID orderId);
 
     @PostMapping("/productCost")
-    Double getPaymentProductCost(@RequestBody OrderDto orderDto);
+    BigDecimal getPaymentProductCost(@RequestBody OrderDto orderDto);
 
     @PostMapping("/failed")
     void paymentFailed(@RequestBody UUID orderId);
